Fix issue #1, improve error reporting
I recently set up SSIS CI for an enterprise and was running into the issue in #1. 
The root cause for the .ispac not being generated was that the project, by default, tries to build in "Interactive Mode" meaning devenv can raise UI dialogs. When testing in PS console this works, but when running the same code via a service (as is the case with the build agent) it fails with a message about Interactive Mode. 
Additionally, error messages from devenv are not returned to the PS session unless you pass "-RedirectStandardError $true" to your Start-Process call, so the call would fail silently.